llamaindex-knowledge-graph/requirements.txt
llama-parse>=0.4.0
llama-index>=0.10.0
llama-index-llms-openai>=0.1.0
llama-index-embeddings-openai>=0.1.0
llama-index-graph-stores-neo4j>=0.2.0
neo4j>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.8.0